[#1127] Add tests for new groovy.test.NotYetImplemented
They requires Groovy 3.0.3 to pass, see:
https://issues.apache.org/jira/browse/GROOVY-9492

Fixes #1127.
