+++ spock-specs/specs.gradle
@@ -11,6 +11,9 @@ configurations {
 dependencies {
   testCompile project(":spock-core")
   testCompile libs.groovySql  //for groovy.sql.Sql
+  testCompile(libs.groovyTest) {  //for groovy.test.NotYetImplemented
+    exclude group: 'junit'
+  }
 
   testRuntime libs.asm
   testRuntime libs.bytebuddy
